refactoring: 휴대전화 확인 페이지 비즈니스 로직  리펙토링
- 유틸함수 분리
- 휴대전화 확인 로직 기능 함수화
