+++ public/javascripts/check_phone_number.js
@@ -1,14 +1,7 @@
-import { getRandomNumber } from './utils/random_number.js';
-import { timer } from './utils/timer.js';
+import { getRandomNumber } from './utils/get_random_number.js';
+import { setTimer } from './utils/set_timer.js';
 
-const phoneNumber = document.querySelector('#phone-number');
-const cancleBtn = document.querySelector('.cancle-btn');
-const check = document.querySelector('.check');
-const phoneNumberForm = document.querySelector('.phone-number-form');
-const authRequestBtn = document.querySelector('.auth-number__request');
-const authReRequestBtn = document.querySelector('.auth-number__request-re');
-
-const getAuthNumber = timer(() => {
+const getAuthNumber = setTimer(() => {
   const authNumber = document.querySelector('#auth-number');
   const nextBtn = document.querySelector('.next-btn');
 
@@ -16,50 +9,80 @@ const getAuthNumber = timer(() => {
   nextBtn.style.display = 'block';
 });
 
-phoneNumber.addEventListener('input', () => {
-  phoneNumber.value = phoneNumber.value.replace(/[^0-9]/g, '');
-  phoneNumber.value = phoneNumber.value.slice(0, 11);
-  phoneNumber.value = phoneNumber.value
+function processPhoneNumberInput(phoneNumber) {
+  let processedPhoneNumber = phoneNumber;
+
+  processedPhoneNumber = processedPhoneNumber.replace(/[^0-9]/g, '');
+  processedPhoneNumber = processedPhoneNumber.slice(0, 11);
+  processedPhoneNumber = processedPhoneNumber
     .replace(/[^0-9]/g, '')
     .replace(/^(\d{0,3})(\d{0,4})(\d{0,4})$/g, '$1-$2-$3')
     .replace(/(\-{1,2})$/g, '');
 
-  if (phoneNumber.value.length === 13) {
+  return processedPhoneNumber;
+}
+
+function setPhoneNumberEvent() {
+  const phoneNumber = document.querySelector('#phone-number');
+  const cancleBtn = document.querySelector('.cancle-btn');
+  const check = document.querySelector('.check');
+
+  phoneNumber.addEventListener('input', () => {
+    console.log('??');
+    phoneNumber.value = processPhoneNumberInput(phoneNumber.value);
+
+    if (phoneNumber.value.length === 13) {
+      cancleBtn.style.display = 'none';
+      check.style.display = 'block';
+
+      return;
+    } else {
+      check.style.display = 'none';
+    }
+
+    if (phoneNumber.value) {
+      cancleBtn.style.display = 'block';
+    }
+  });
+
+  cancleBtn.addEventListener('click', () => {
+    phoneNumber.value = '';
     cancleBtn.style.display = 'none';
-    check.style.display = 'block';
+  });
+}
 
-    return;
-  } else {
-    check.style.display = 'none';
-  }
+function setAuthNumberEvent() {
+  const phoneNumber = document.querySelector('#phone-number');
+  const authRequestBtn = document.querySelector('.auth-number__request');
+  const authReRequestBtn = document.querySelector('.auth-number__request-re');
 
-  if (phoneNumber.value) {
-    cancleBtn.style.display = 'block';
-  }
-});
+  authRequestBtn.addEventListener('click', async () => {
+    const authNumber = document.querySelector('.auth-number');
 
-cancleBtn.addEventListener('click', () => {
-  phoneNumber.value = '';
-  cancleBtn.style.display = 'none';
-});
+    if (phoneNumber.value.length !== 13) {
+      return;
+    }
 
-authRequestBtn.addEventListener('click', async () => {
-  const authNumber = document.querySelector('.auth-number');
+    phoneNumber.readOnly = true;
+    authRequestBtn.style.display = 'none';
+    authNumber.style.display = 'flex';
+    getAuthNumber();
+  });
 
-  if (phoneNumber.value.length !== 13) {
-    return;
-  }
+  authReRequestBtn.addEventListener('click', () => {
+    getAuthNumber();
+  });
+}
 
-  phoneNumber.readOnly = true;
-  authRequestBtn.style.display = 'none';
-  authNumber.style.display = 'flex';
-  getAuthNumber();
-});
+function initCheckPhoneNumberPage() {
+  const phoneNumberForm = document.querySelector('.check-phone-number__form');
 
-phoneNumberForm.addEventListener('submit', (e) => {
-  e.preventDefault();
-});
+  setAuthNumberEvent();
+  setPhoneNumberEvent();
 
-authReRequestBtn.addEventListener('click', () => {
-  getAuthNumber();
-});
+  phoneNumberForm.addEventListener('submit', (e) => {
+    e.preventDefault();
+  });
+}
+
+initCheckPhoneNumberPage();
