+++ public/javascripts/utils/get_random_number.js
