+++ public/javascripts/utils/set_timer.js
@@ -1,4 +1,4 @@
-export function timer(callback, time = 2000) {
+export function setTimer(callback, time = 2000) {
   let timeout;
 
   return (...args) => {
